requirements.txt
# Clinical Document Intelligence Platform
# Core dependencies

# === API Framework ===
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4

# === Frontend ===
streamlit==1.43.2
requests==2.32.3

# === LLM & Embeddings ===
httpx==0.28.1
tiktoken==0.8.0

# === Vector Store ===
chromadb==0.5.23

# === Data Processing ===
pyyaml==6.0.2
jinja2==3.1.4
python-dotenv==1.0.1

# === Testing ===
pytest==8.3.4
pytest-asyncio==0.24.0

# === Optional: Re-ranking (install separately if GPU available) ===
# sentence-transformers==3.3.1

# === Optional: AWS Bedrock ===
# boto3==1.35.86

# === Optional: Redis Caching ===
# redis==5.2.1

# === Optional: Monitoring ===
# prometheus-client==0.21.1
